Imported Upstream version 2.8.5.2

+++ src/dynamic-preprocessors/ftptelnet/hi_util_kmap.c
@@ -237,7 +237,7 @@ int KMapAdd( KMAP *km, void * key, int n, void * userdata )
     if( n <= 0 )
     {
         n = strlen( (char*) key );
-        if( n > sizeof(xkey) )
+        if( n > (int)sizeof(xkey) )
             return -99;
     }
     
@@ -380,7 +380,7 @@ void *  KMapFind( KMAP * ks, void * key, int n )
     if( n <= 0 )
     {
         n = strlen( (char*)key );
-        if( n > sizeof(xkey) )
+        if( n > (int)sizeof(xkey) )
             return 0;
         
     }
